refactor: native classes and interfaces
